fix: cast cross-enum comparison in basic_test to suppress -Wextra warning
Agent-Logs-Url: https://github.com/corepunch/orion-ui/sessions/5dbde6e7-47db-41f4-8150-c57ee7cfff05

+++ tests/basic_test.c
@@ -84,7 +84,7 @@ void test_control_message_constants(void) {
     TEST("Control message constants");
     
     // Verify control message constants
-    ASSERT_TRUE(btnSetCheck >= evUser);
+    ASSERT_TRUE((int)btnSetCheck >= (int)evUser);
     ASSERT_TRUE(btnGetCheck > btnSetCheck);
     ASSERT_TRUE(cbAddString > 0);
     ASSERT_TRUE(cbGetCurrentSelection > 0);
